requirement.txt
fastapi 
uvicorn
sqlalchemy 
psycopg2-binary 
pydantic 
python-dotenv
databases
asyncpg
alembic
pytest
httpx
pytest-asyncio
passlib[bcrypt]
python-jose
email-validator
regex